ADD|UPDATE - ViewModel untuk mengelola status UI dan operasi terkait event, termasuk pengambilan dan penghapusan data event. ganti type data menjadi integer sesuai dengan database

+++ app/src/main/java/com/example/eventorganizer/model/Events.kt
@@ -19,7 +19,7 @@ data class EventsDetailResponse (
 @Serializable
 data class Events (
     @SerialName("id_event")
-    val idEvent: Long,
+    val idEvent: Int,
 
     @SerialName("nama_event")
     val namaEvent: String,
